requirements.txt
python-binance
requests